backend/requirements.txt
asgiref==3.2.10
attrs==20.2.0
certifi==2020.6.20
chardet==3.0.4
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
idna==2.10
iniconfig==1.0.1
packaging==20.4
pluggy==0.13.1
psycopg2==2.8.6
py==1.9.0
pyparsing==2.4.7
pytest==6.1.1
pytest-django==3.10.0
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
